+++ examples/plot_gong.py
@@ -15,7 +15,8 @@ import numpy as np
 import sunpy.map
 
 import pfsspy
-import pfsspy.coords as coords
+from pfsspy import coords
+from pfsspy import tracing
 
 
 ###############################################################################
@@ -92,17 +93,24 @@ ax.set_title('Source surface magnetic field')
 fig = plt.figure()
 ax = fig.add_subplot(111, projection='3d')
 
+tracer = tracing.PythonTracer()
 # Loop through 16 values in theta and 16 values in phi
 r = 1.01
-for theta in np.linspace(0, np.pi, 17):
-    for phi in np.linspace(0, 2 * np.pi, 17):
-        x0 = np.array(coords.sph2cart(r, theta, phi))
-        field_line = output.trace(x0)
-        color = {0: 'black', -1: 'tab:blue', 1: 'tab:red'}.get(field_line.polarity)
-        ax.plot(field_line.x / const.R_sun,
-                field_line.y / const.R_sun,
-                field_line.z / const.R_sun,
-                color=color, linewidth=1)
+theta = np.linspace(0, np.pi, 33)
+phi = np.linspace(0, 2 * np.pi, 33)
+theta, phi = np.meshgrid(theta, phi)
+theta, phi = theta.ravel(), phi.ravel()
+
+seeds = np.array(coords.sph2cart(r, theta, phi)).T
+
+field_lines = tracer.trace(seeds, output)
+
+for field_line in field_lines:
+    color = {0: 'black', -1: 'tab:blue', 1: 'tab:red'}.get(field_line.polarity)
+    ax.plot(field_line.x / const.R_sun,
+            field_line.y / const.R_sun,
+            field_line.z / const.R_sun,
+            color=color, linewidth=1)
 
 ax.set_title('PFSS solution')
 plt.show()